replaced external task client bean
