+++ charts/civil-damages-service/Chart.yaml
@@ -2,7 +2,7 @@ apiVersion: v2
 description: A Helm chart for civil-damages-service App
 name: civil-damages-service
 home: https://github.com/hmcts/civil-damages-service
-version: 0.0.6
+version: 0.0.5
 maintainers:
   - name: HMCTS Unspecified team
