+++ charts/civil-damages-service/values.aat.template.yaml
@@ -57,7 +57,6 @@ java:
     FEIGN_CLIENT_CONFIG_REMOTERUNTIMESERVICE_URL: http://civil-damages-service-camunda-staging-aat.service.core-compute-aat.internal/engine-rest/
     FEIGN_CLIENT_CONFIG_REMOTEEXTERNALTASKSERVICE_URL: http://civil-damages-service-camunda-staging-aat.service.core-compute-aat.internal/engine-rest/
     FEIGN_CLIENT_CONFIG_REMOTEREPOSITORYSERVICE_URL: http://civil-damages-service-camunda-staging-aat.service.core-compute-aat.internal/engine-rest/
-    CAMUNDA_BPMN_CLIENT_BASE_URL: http://civil-damages-service-camunda-staging-aat.service.core-compute-aat.internal/engine-rest/
     TESTING_SUPPORT_ENABLED: true
     POLLING_EVENT_EMITTER_ENABLED: false
     LAUNCH_DARKLY_ENV: service-staging
