+++ src/main/java/uk/gov/hmcts/reform/unspec/handler/tasks/StartBusinessProcessTaskHandler.java
@@ -7,7 +7,7 @@ import org.camunda.bpm.client.task.ExternalTask;
 import org.camunda.bpm.engine.delegate.BpmnError;
 import org.camunda.bpm.engine.variable.VariableMap;
 import org.camunda.bpm.engine.variable.Variables;
-import org.springframework.context.annotation.Configuration;
+import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDataContent;
 import uk.gov.hmcts.reform.ccd.client.model.Event;
 import uk.gov.hmcts.reform.ccd.client.model.StartEventResponse;
@@ -21,7 +21,7 @@ import uk.gov.hmcts.reform.unspec.service.flowstate.StateFlowEngine;
 
 import java.util.Map;
 
-@Configuration
+@Component
 @RequiredArgsConstructor
 @ExternalTaskSubscription(topicName = "START_BUSINESS_PROCESS")
 public class StartBusinessProcessTaskHandler implements BaseExternalTaskHandler {
