+++ src/main/resources/application.yaml
@@ -135,10 +135,6 @@ feign:
       remoteExternalTaskService:
         url: http://localhost:9404/engine-rest/
 
-camunda.bpm.client:
-  base-url: http://localhost:9404/engine-rest/
-  async-response-timeout: 29000
-
 fees:
   api:
     url: http://localhost:8765
